make classes public
